+++ CHANGELOG.markdown
@@ -1,3 +1,11 @@
+next [????.??.??]
+-----------------
+* Fix a bug in which `deriveBifunctor` would fail on sufficiently complex uses
+  of rank-n types in constructor fields.
+* Fix a bug in which `deriveBiunctor` and related functions would needlessly
+  reject data types whose two last type parameters appear as oversaturated
+  arguments to a type family.
+
 5.5.7 [2020.01.29]
 ------------------
 * Add `Data.Bifunctor.Biap`.